+++ assert.js
@@ -1,1 +1,0 @@
-build/assert/index.js